Prevented loading images in ListView.onMeasure()
The problem is that onMeasure() reloads the first few rows
very often for measurement purposes.  When it loads a row,
the ImageCache was trying to load the URL.  This caused a lot
of started/cancelled downloads, bogging down the system.

Now I've got a custom ListView which will inform the adapter
if it's just measuring.  If it is, then the adapter will
not do as much fancy processing.