7118: Fixed invalid item id saving
When the map tab was visible and you filtered the items,
it wouldn't properly requery the data on the ListView. This
meant that when the ListView tried to save its state it
could end up trying to save data that no longer existed,
crashing the app.

This should be enough to fix it, but there may need a more
complex fix later.