6670: Imported customized version of Gallery
This will allow us to get fine-grained control over Gallery behavior
and fix some issues.

I also started the work of privatizing an Adapter just for the Gallery,
by moving ImageAdapter inside of it.  This allows the Adapter to be
coupled to the Gallery, which normally would be a Bad Thing but in this
case may be necessary to get the exact behavior we want.