6670: Fixed gallery offset resets
What happened before was that the gallery would "reset" to
center whenever a new image is loaded.  This is because the
layout() method had no concept of where the current offset
is.  Now it'll take into account the current offset from
the selected item, and make sure that that offset is initially
set when doing a new layout.