8410: Fixed bug with duplicate URLs
With trimming, if the url list had duplicates, it would be possible
to trim the URL of an image that is also being displayed at that
very moment.  For example, if an image was in slot 5 and slot 25,
when you viewed slot 5 it would clear it due to being so far away
also in slot 25.  This would cause a crash.

Now we deduplicate the URLs so that this cannot happen.