wdgt: getting rid of searchResponse from state
The reason for this is because we don't actually need
to hold onto the searchResponse for each widget. The
max of 5 properties that pertain to the widget
are held on to as another member variable. Holding on
to the searchResponse takes up a lot of memory that isn't
really used.