Merged HotelLib into ExpediaBookings
This is a merge of an entire lib submodule in the main project.  As a result,
history has been lost for these files.  If you need to know their history,
find the file in HotelLib (which is no longer required for builds).

In case people are reading this years from now: back in the day, we used
to have three hotel-related applications: HotelPal, StayHIP, and Expedia
Hotels.  We used one library between the three, HotelLib.  As time went on,
we stopped supporting HP and SH.  As a result, it was unnecessary work to
keep maintaining HotelLib as a separate project.  It also bloated the
build size.