8774 and 8780: fix scroll bar
These fixes resulted in the rewrite of quite a bit of the
drawIndicator helper method. This is mainly due to the
assumption thus far of every row being of the same height.
The fix essentially accounts for headers in the list view
while calculating the scroll size, the position of the indicator
in the scroll bar and the nmber of visible items in the list.

A couple of working assumptions:
a) the height of each row item is the same
b) the height of each header item is the same (while I do calculate
the total height occupied by the header based on the height of each
header row, I have a mHeaderHeight variable that i use in places
for calculations when all header rows are not visible)