9336 and 9596: improxing widget UX
tapping the widget now starts the search
activity instead of the hotel activity and updates
the searchParams of the searchActivity to match that of the widget
that was just clicked. This helps to ensure that hitting back
on the hotel details screen will take you to the search from which
the widget originated. While the results are not cached, the search
is just re-downloaded.

Another change is to have the branding cell take you to the search
activity to present you with the results in the widget