9593: making sort popup bigger
and also fixing bug with how the
sort popup was being dismissed. Instead
of setting the outsideTouchable flag
to be true in the popup, using a transparent
view which upon being clicked dismisses the popup.
The reason for this is to be able to click
any of the sort options when the popup is being shown
and avoid weirdness in how it is dismissed if any of the
elements in the popup is clicked