9705: fix for widget to pickup filter
the widget so far was ignoring any filtering
that was happening in the app. Now, the widget
first applies the filter to the results downloaded
and tries to fill all slots with properties from the filtered
list. If there aren't enough, then all properties are traversed
to fill slots for the widget.