9712: one time widget notification
widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown
9712: one time widget notification

widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown
9712: one time widget notification

widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown
9712: one time widget notification

widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown
9712: one time widget notification

widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown
9712: one time widget notification

widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown
9712: one time widget notification

widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown
9712: one time widget notification

widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown
9712: one time widget notification

widget notification is shown after 3 app
searches on a new or upgraded install.
If there are widgets installed, the
notification is not shown