Fixed up showing distance/exact location
The previous system was kind of jumbled mess that had a few bugs
in it.  It wouldn't necessarily show exact location all the time, and
it kept contradicting itself (setting to show distance then immediately
thereafter disabling it).

The new system should be more clear.  You either need to use
determineWhetherExactLocationSpecified() to let the system figure out
what should be set, or you should specifically yourself what is shown
via setShowDistance() and showExactLocation() (or dontShowExactLocation())