9780: Distance filters hidden sometimes
When user searches for results which do not yield distances (e.g., too
general like "Boston, MA") then we will now hide the filter-by-search-radius
options now.

Made sure to set search radius to ALL in this circumstance, so no results are
hidden from the user due to use defaulting to LARGE most of the time.