more details from mini details fragment
clicking on the see details button opens up
the details fragment. When the user then navigates through
the list, both mini details and full details fragments
are updated with the content.
Things still dont work on rotation, but this is a good point
to push what I have so that we can see things functional