hotel image collage on mini details view
refactored the code to create
a HotelCollageHandler that manages the
updating and the clearing of the image views
in the collage. This is so that the code
can be re-used in the mini details fragment.

Ideally, I would've liked for the layout
code to stay the same for the mini details
fragment and the hotel details fragment. Since
the layout of the collage is different and this code will
not be used in any other place, I just created two different
layouts for the views for now, as the work to dynamically
change the layout in code is more complex than is warranted