Fixed crash
When you do a search with results, then do another search that has no results,
it was crashing because the scrollbar was trying to measure itself.  This seems
to be a caching issue; the adapter still shows the header, but there are no
items drawn into the listview until its shown to have results.