10211: Removed leaky MyLocationOverlay
MyLocationOverlay is apparently causing massive memory leaks, such that
if you start exiting/entering the app it will cause a crash (if you're
always viewing the map).  I've thus removed it from the app for now.

In its place, an ExactLocationOverlay is now used.  It is not always 100%
up-to-date (if the user doesn't do a new search but moves, for example)
but it's better than having no indicator of current location.