10563: No search params on availability screen
Changed how we determine whether or not
to show the search parameters in the actionbar.
Before this change, the condition was to ensure
that there was nothing in the backstack. With
the new requirement of not showing the search
parameters when the availability list is up,
this assumption is no longer valid.

Now, we do not show the actionbar if any of the following
conditions are met:
a) User is on the search screen
b) User is on the availability list screen
c) User is on the confirmation screen