ensure that there is just 1 balloon open
it was possible for multiple balloons to
show on the map after navigating back
from the details screen to the hotel map.

This change fixes that by ensuring that no overlay
is held on to after destroyView is called
on the HotelMapFragment.

The problem was that a stale
hotelOverlay was having its state
updated while the HotelMapfragment
was in a state where the fragment was
removed but being managed by the
fragmentManager (when the hotel details
are showing)