10526: postal code hint fits
made the width occupied by the postal
code box dynamic such that it occupies
a different width on all devices prior to
xlarge-mdpi.

This enables us to ensure that the postal
code hint fits in the box, but the caveat
here is that the space occupied by the
country spinner is no reduced due to the
limited space in potrait mode. What this means
is that there is a higher probability
of the country not fitting in the spinner.
This is inevitable, just a matter of how
often the country fits or doesn't fit.