10587: loading reviews into details
Setting aside a set amount of space for
displaying reviews. The amount of space
set aside is dependent on the number of reviews
that can be shown. If just one row of reviews
are available (up to 2 reviews) half the
amount of space is set aside than if there
were more than 2 reviews that could be displayed