Removed URL HashMap from Media
This means that URLs are now created when they are needed, instead of
having all the overhead during parsing to create the alternative URLs.

Also simplified access to high-res images.  Classes no longer need to
know to start at the highest-res image - just pass in an ImageView and
a callback to start downloading a high-res image into that ImageView.