11058: Changed behavior of property selection
The behavior change is that whenever a property is selected, it will advance
the user - regardless of where the click came from.  By "advance", I mean that
it will either change the property selected, or expand the next fragment (in
the case that the property was already selected).

This both fixes a bug that occurs, and also gives users multiple avenues
for progressing in the app that didn't exist before.