11181: Stopped clicks fallthrough mini-dts
The mini details is being placed on top of the map.  As a result, clicks on
the details which weren't being handled were being taken in by the map,
allowing people to interact with the map behind the the mini details.  Now
that is prevented from happening by capturing all clicks on the mini details.