fixing room rates calculation
when the minimum rate available
was the only rate for all bed types, the
summarized room rates container would not have
any rate mappings added to it since all rates
were considered the minimum. The fix here is to
ensure to avoid picking just one bed type mapping
that is considered to be the starting rate versus all bed type
mappings for the minimum rate.