11403: Massaged data for Time parser
It was blowing up on API <= 7 due to a bug in their parser.  Supposing that
the format we get from Expedia is always the same, I attempt to massage the
data into a format that doesn't blow up.

I've made sure to only massage the data when we know (for sure) that it's
off.  Otherwise, it will use the original data (hoping that it can be
parsed correctly).