Replace parseDescription with fastParseDescription
More notes are inline the code but tests show this is up to 20x faster
than the old implementation with regexs which makes sense since we were
walking the string over and over again. A little more work to do to make
things properly fast but a good first attempt.