12608: Fixed bug in getDaysBetween()
It's a timezone bug.  Imagine you're comparing two dates - 10 PM Central
time on DAY, and midnight GMT on DAY + 1.  The "day of year" of the second date
is higher than the "day of year" on the first.  HOWEVER, the actual time
represented by the first date is *after* the second date (due to timezones).

As a result of this logical mishap, getDaysBetween() would reverse the order
of steps, and as a result would end up in an infinite loop.

Now it checks the days between based on Calendar fields, rather than using
the after() method (which uses pure millisecond time).

(cherry picked from commit 80cdb0bfe7e810f218a272972437a96a4facf3f8)