12812: Add something after RulesRestrictions link
Fixed a few languages. Most have a period after the link or some kind of
text. Without it the anchor tag of the privacy link extends to the end
of the TextView. So when the legal link and privacy link are on
different lines but NOT on top of one another you end up hitting the
privacy link 1/2 the time.