Fix serious issue with activity reference
During the refactor of encapsulating the bookkeeping on the different tabs,
I created an object TabSort that collected and stored this data. Additionally,
I moved the Download and OnDownloadComplete object to be an inner class and object
of this TabSort object.

The problem is that I stored the TabSort object across orientation changes, but as a
non-static inner class, it held implicit reference to the previous Activity instance.
After orientation change, it went ahead and referred to old object from the old Activity
instance.

In order to fix, I had to change the TabSort to be a static inner class and then send an
activity object in the constructor so that it could access activity objects. Thanks Scott!!