12830: Prevent unnecessary concurrent widget search requests
For some reason, when the widget was starting up making 8+ requests
at a time.  I've gone through and patched it so it only makes one
request at a time when it starts up.

This commit is a band-aid.  It doesn't fix the logic underlying the
aggregious requests, but it prevents them from damaging the network
by sending so many requests at once.