Update user reviews list to include a divider
This is strictly for the recent sort order in the event that there are reviews
segregated by language. It simply displays a divider/header/title stating
"Reviews in other languages"

It is not certain that this feature will be in the next release, but we will go
ahead and get the string localized at any rate.