12681 better error handling on reviews activity
This commit made many changes to the error-handling code for the reviews activity,
essentially beefing up the error-handling support. The main change is that there has
been introduced a new status/error message that gets displayed in the event that the
response by BazaarVoice contains error OR if there were errors parsing the response.

Here is the expected behavior: If the review statistics call returns with errors, or we
have trouble parsing the response, display "Reviews are temporarily unavailable" for all
three review sorts. Additionally, do not populate the number of reviews + rating bar in the
bottom bar, although, display the Book Now button (hey, we still want the user to be able to
book from this page even though reviews don't work :P). If the reviews call fails, display
the message "reviews are temp unavailable" for the given review sort.

There does not need to be as informative errors as the E3 related stuff, because there
is no human error that can be introduced (the user is not able to construct data that
is sent as part of the request). In that way, error handling does not need to be nearly
as robust. This is a step in the right direction and may be sufficient, although there
are a whole host of error codes that could potentially be returned by BazaarVoice if we
for some reason have the desire to do so at a later time.