12826: Fix race starting 2 user review activities
Click on a hotel with no loaded information and quickly tap the reviews
button as fast as possible. I am able to do it with ease on my Galaxy
Nexus. We were starting 2 User Reviews activities, the first would load
the reviews just fine but when you hit back the second would appear and
the downloads would be canceled which ends up showing a progress bar
forever.