13062: Workaround for EditText on ICS
In ICS and beyond, EditText's have an entire new action mode for editing.
As a result, selectAll() doesn't work (and it drastically changes the UI
to properly activate the selection mode).  This is an interim fix that at
least starts the cursor at the END of the input rather than the start to
make it easier to delete the text.