13072: Show distances more liberally than before
Before we only showed distances (and related widgets) if the search
was for an exact address (or current location).  With the addition
of Expedia's autosuggest, this no longer quite makes sense -
sometimes a POI will have a lat/lng and you want to show distance to
that POI.

Since we don't have time for more advanced logic, for the time being
both iOS and Android are just liberally showing distance from a
search point.