13048 stop bad bug for now
There is some messy code in the background and it leads to mixing up reviews in tabs
when the separate tabs have been chained to download. Not sure what is going on exactly
but in interest in getting this release out, disable the  chaining so that this bug does
not in fact occur.