Initial commit of UserReviews refactor
Refactor this activity to be compatability libarary fragment based. This
will hopefully drastically reduce the amount of code required to coordinate
the three different sorts, and make the code cleaner and more maintainable.

This initial commit is a robust implementation of the activity with the option
of only one sort (no switching). Additionally, reviews are stored in a singleton
class for the interim, until a more suitable long-term solution is figured out.

Next steps: tabs + use Db