Offers data propogation/reuse
Now the HotelActivity will use an avail response if it is already in the Db.
Additionally, it will get this info back from future requests if it is there.

It also means that it will save offers requests between clicks, so it doesn't
need to reload the page each time you click on a hotel.