13554: Fixed NPE on rotation
If you rotated the screen *after* the search results have loaded, then
you'd run into a crash.  The problem was that it was assumed that
onCreateOptionsMenu() would be called after onCreate().  That usually
happens, but we were calling invalidateOptionsMenu() during onCreate()
if the search results had already been created.

Now we check to see if onCreateOptionsMenu() has been called before
allowing invalidateOptionsMenu() to be used.