Switched to more advanced flight search params
The old setup made it difficult to do multi-leg flight searches.  The
new format will make it simple to create multi-leg flights (eventually).

There still remains a simple interface to FlightSearchParams that
assumes that the search is either one-way or round trip.  But it should
only be used for these two circumstances (if and when we implement a
more complex, multileg search we will need to abandon these).