13365: Quit activities on expired db
In the old days, if the memory expired, the app would launch into
the last Intent and still show data (but when you backed out, it
would have to reload all its state).  Now that we use a central Db,
we need to expire and kick the user back to the beginning in this
situation.

We auto-finish activities when they have expired data now.  This
should kick users back to the beginning of the app, with a new
search starting automatically.