13586: Fixed one-night searches on 2.1/2.2
This bug was being caused by a very obscure Android issue.  Here are
the conditions under which the problem occurs:

1. You have two Calendar objects that are one day apart.
2. You have one SimpleDateFormat that you want to use.
3. You set both the Calendars and the DateFormat to the same TimeZone.
4. You format them both, one after another.

In this particular sequence of events, the formatted output is the same.
It seems that the SimpleDateFormat has some state bug in it that keeps
it from properly formatting the times in this circumstance.

The solution here is to simply create another DateFormat, so that
the state bug doesn't happen in the first place.

Note that SimpleDateFormat works properly on 2.3+.