13715: Properly cancel requests on Activity exit
Right now, the request for info would actually survive past the
Activity's lifecycle.  This meant that not only was there a retained
copy of the Activity in memory, but also that you could re-enter that
Activity from another hotel and get its results (instead of loading
new ones relevant to the current hotel).