13786: Use our own domain validation for cookies
Create our own CookieSpecRegistry and store our own cookie spec handler
inside. The only place where the NetscapeDraftSpec cookie spec was
failing was for POS's not in the `special domain' set that had less than
3 parts. So ".expedia.ca" or ".expedia.de" has 2 parts but
".expedia.co.jp" has 3 and "expedia.com" is in the special list so it is
allowed to have only 2 parts.