13780: Don't use all invalid field errors from server
When using a stored credit card and the wrong CVV the server says that
the Credit Card Number (not displayed) and the expiration date (not
displayed) are invalid. However these are not valid fields for a request
with a stored credit card. So when using a stored credit card we filter
out any error invalid field error returned from the that is non-sense.