Updated FlightAdapter to support expandable rows
There was quite a refactor I had to do in order to get this to work.
Now FlightListFragment itself contains the FlightAdapter, and the
adapter handles expanding/shrinking rows.  In addition, the listeners
for clicks have been modified to better represent the situations that
can now present themselves (like details vs. select).