Moved loading flights into FlightSearchActivity
It used to load in FlightSearchResultsActivity.  However, this presented
some problems: in particular, if you are doing a multi-leg flight, it
doesn't make sense to keep reloading the flights.  The loading now
occurs on the search page, giving you an option to quickly fix your
search parameters if you messed up on the initial search.