ListView animation slides instead of scales
Before, it was just scaling the size of the expanded details container.
This had the effect of making the expanded details container look kind
of silly (as all its Views would kind of go wonky during the animation).
Now, with an animation View that does the scaling (but the details
staying the same size) the animation looks a lot cleaner.