Fixed up FlightSearchParams
It was misreporting its number of query legs.  This was due to changes
made previously to adapt to the new CalendarDatePicker behavior.

I've reverted some of the code so that it works more like it did before -
maintaining the correct number of legs for the query, and maintaining
its round-trip status by inference rather than an explicit variable.