Add basic end-to-end test through Flights flow
Test starts at FlightSearchActivity and ends at the FlightDetails screen. This is
not a useful test yet, but it is useful in that it has uncovered the tecniques for
cycling through the application through a UI test.

Also add some basic package organization for the test project. Now includes activity,
tests, and utils sub-folders as not all files in the test project are test cases.