Added field validation visual feedback.
There is now a new SectionFieldValidIndicator type that acts as validation feedback, e.g. it turns green when the validator passes. It is built like the other SectionFields and gets bound in the same way. We also refactored the validation scheme a little bit, it should now be much faster because if the value didn't change we just return a stored boolean.