Removed expandable list view functionality
The new designs no longer call for the list rows to expand/contract.

Let's have a moment of silence for this awesome, awesome code which must
now go to the grave.  It was truly taken before its time.  It showed me
how great ListView animations could be and how freaking complicated a simple
feature can be to code.