Add one more layer of organization.
This way, tests can be more easily invoked by group. The plan is to have unit
tests run on every build, adn the UI, integration tests run not as often. Using
instrumentation, tests by package can be invoked. There may be another way to
achieve the same sort of behavior (perhaps by attaching an annotation to a test),
although, being broken down in this manner is straightforward for readability.