114: Search/SearchParams SearchType revamp
SearchType was very under-used and confusing. Basically only FREEFORM and
MY_LOCATION were being used, and the few other enum values were not.
I added a few enum values: CITY, ADDRESS, POI, and saved the SearchType
in the previous searches with ActiveAndroid (Search.java).

In the meantime, the property "FreeformLocation" was being used for more than
just that, so I renamed it to a more proper "Query" to show that it will be
used for the query string, no matter what the SearchType is.

The bottom line is that we'll now keep track more accurately of the SearchType
so that we can use it when we want to (like to determine whether distance
or neighborhood should be displayed on the Search Cell).