402 - Added the navigation dropdown.
The navigation dropdown is not yet fully functional - some of the pages that it is supposed to point to do not yet exist.
It did not get added to some pages (or any of the hotels activities). It would be nice to have a more unified way of adding it, instead of going through and touching every single activity.
Finally as I was in touching various activities and updating the Actionbar text, I went in and restyled just a little bit.