231: Simplify deals sort, and correct it.
The comparator was sufficiently complicated that it is easier to filter
the properties into two lists. This also allowed getting the cost of the
comparator down as well since we now only pass deals that are on sale to
it. Now the code is explicit in what it is doing and not hidden behind
some complicated semantics inside of a comparator.