Prevent memory leak in PlaneWindowView
Running a suite of integration tests exposed an OutOfMemory error within
PlaneWindowView, although, it can be reproduced by simply entering and exiting
the FlightSearchResultsActivity on repeat (specifically StatusFragment hosting
PlaneWindowView).

The SensorEventListener was not being unregistered before the thread was destroyed,
thus the SensorManager held reference to every instance of PlaneThread (and its
expensive Bitmaps).

Discovered culprit using the hprof dump and MAT in Eclipse