Add more FlightSearchParamsTests
The new tests assert search params from UI to search params in the request URL.
It should be noted that these tests retrieve the "actual" (when asserting
expected vs. actual) from debug logcat logging from ExpediaServices. Ideally,
these actual values will come from snooping the network requests. The tests
are written in a way such that it will be easy to rewrite and improve on the
the way in which the "actual" values are returned.