Updated how endpoint URL is calculated
Instead of automatically generating part of the URL (either
MobileHotel/Webapp/ or api/flight/), we just put that in each request.

The reason for this change is because the endpoints keep moving about
in the API these days and it no longer makes sense to limit things
to one endpoint or another.

Also formatted an enum.