664: Prevented PlaneWindowView loading twice
The problem had to do with the split action bar.  If the PlaneWindowView was
added before onCreateOptionsMenu(), then it would have to reload itself once
the Activity turned into split action bar mode.

This is now prevented by starting the search/loading fragment after the menu
is created (instead of before).