Renamed "passenger" to "traveler"
I decided to keep the variable names consistent when renaming FlightPassenger
to Traveler.  Otherwise I think that the naming of params could get a bit
confusing for someone stumbling on the scene after the fact.

If you see this in git blame you can pretty safely ignore it because it's just
a variable name change.

This is part of a larger change to make a generic Traveler type that handles
both flights and hotels.