f439 Add Flights Omniture tracking class
Basically, add an enormous class that attempts to cleanly convert the spec
into methods using the Omniture jar file that we have. Spec:
http://confluence/display/Omniture/Mobile+App+Flight+Tracking

The premise is to make public methods that can be added into our business
classes as inobtrusively as possible. The methods rely on Db.java for most
things such that the tracking methods can be plugged into business code easily,
without having to search for parameters. Hopefully this reduces the impact of
future changes and keeps the tracking logic separated from business logic as
much as possible.