Move Kevin's Hotels test to a new package
There has been interest expressed in having a separate Jenkins job that run
the tests that Kevin is writing for Expedia Hotels. We are attempting to
figure out a more scalable solution than creating a new package every time
that a new Jenkins job is required, but this is the quick and easy way for
the time being.