H546: Fix deals sort comparator
TimSort is able to detect the following violation of the comparator
contract that I had hoped it would not detect. In particular we we're
not reporting that two items were ever equal in the DEALS_COMPARATOR
because if the discount percentages were equal I wanted to maintain the
exact same order as the popularity sort. To fix the crash I am just
reporting equal discounts as equal.