Apparently there exists IntentCompat
This should resolve the backstack issue for 2.1+. I have yet to test it
on an actual device. Does not work on the emulator, although, emulators have
historically had Intent issues.

Also, it seems a little bit like black magic to use IntentCompat.FLAG_* as a
flag to the vanilla Intent, although, the API does not lead me to believe it
should be used in a different way.. Perhaps I have a misunderstanding?