Fix some flow issues regarding launch screen
The problem is that we have a launch screen and also a navigation dropdown that
allows the user to navigate to the launch screen (and hotels and other stuff)
from deep within the activity stack. Properly managing this stack on 2.x and
above requires a little hack in our case.

The routing activity (SearchActivity.java) seems to throw a wrench in using
activity intent flags to properly clear the history backstack with (CLEAR_TOP).
CLEAR_TASK seems to work but was introduced in 3.0.

This hack introduces ActivityKillReceiver, a little helper class that is essentially
a BroadcastReceiver with a some public methods to provide easy integration into our
activities. Send the KILL_ACTIVITY broadcast any time that you would like to destroy
the activity backstack. Activities must opt-in by integrating this class.