More 2.0 flow fixes
The two different paths (Flights/Hotels) have "sticky" confirmation pages. This
means that whenever the user attempts to access Flights or Hotels from (a) the
nav dropdown (b) the header from the launch screen AND the user just succesfully
booked, show the user the confirmation page.

Additionally, SearchActivity automatically routed the user to the EH confirmation
screen. We now have a pretty launch screen, so send the user to the launch screen
in this case and only send them to conf screen if the user explicitly attempts
to launch the path that has a recent confirmation saved.

EH tablet is special case. Ensure the app launches to the correct screen in the
event of a sticky conf/booking screen on tablet