Refactor navigation to launch screen
Now that we have a proper launch screen that we intend to keep in the backstack
and return to, we do not need to resort to the KILL_ACTIVITY receiver to navigate
back to the launch screen. Setting the LaunchActivity launchMode="singleTop" and
starting the launch activity with CLEAR_TOP | SINGLE_TOP suffices.

The only potential problem I foresee is if for some reason the launch activity
gets flushed out in interest of low memory, the backstack may not properly clear
(this is what the docs make it seem like, anyway).