Fix navigation pertaining to EH confirmation page
We want the launch activity to stay in the activity stack when the user
navigates back from EH confirmation screen. CLEAR_TASK flag will remove the
launch activity, so remove this usage.

Removes the CLEAR_TASK flag on ConfirmationFragmentActivity start and also
removes the limit on activity finishing hack such that it will finish()
activities if there exists saved confirmation data for all versions
of Android to behave as expected.

This also unblocks Kevin on automated hotel search tests.

I wonder if design wants the up arrow on the ConfirmationActivity?