Relaunch animation on launch onResume (f967)
We are having a problem with attempting to reuse ImageCache bitmaps that are
being recycled elsewhere in the application. This fix simply resets the adapters
to reinitiate the launch animation. Perhaps not ideal, but we are strapped for
time and this is rather simple.

Also, as a side effect of only maintaining one set of Hotel search results in
Db.java (and now that the launch screen does not blow up when returned to),
the set of search results that are set (perhaps when performing a new search in
PhoneSearchActivity) are reflected in the Launch screen. I could kick off a new
hotel search here, although, we probably want to return the user to a screen
without much latency. Must ask product what to do here.

This resolves f967 and h654, essentially duplicate bugs.