F697: Fixed FadingImageView drawing on 2.x
It was a mistake to depend on Canvas.getClipBounds() to determine where to
draw the fade.  It turns out that in some situations, the clip bounds move
around (as the View determines what should/should not be drawn).  So now
it uses onLayout() to determine the bounds and reuses those throughout.