f1011 don't let user leave app near checkout
Our t&c, privacy policy, legal information links provide such a great experience
for the user on a phone we want to make sure that the user has this experience in
shitty WebView rather than their browser.

But really, we don't want to send the user outside of the app from the checkout
screen.