F1073: Fixed recovery on search screen
We need to be careful how we handle this error.  Normally we can determine
whether or not to restore the instance state based on null data, but in this
case we are expecting the search params to perhaps only be partially full
on entrance.

What we do instead is detect the situation where it *must* be full (during
an update on resume) and restore the state there if we didn't earlier.