Improve launch screen edge case
Launch screen did not work well in the event of hotel search failing to return
results. Specifically, search seems to fail when searching for a day near the
end of the day. In the event of failed search, increment the checkin/checkout
dates and set the location to somewhere in SF.