F1063: Use API carrier names instead of local FS.db
We should always use Db.getAirline() now, as that will properly
fill in the airline name data.

There is an assumption made in the code that there is only one name
for any given airline.  It turns out that the normal airline name is
different from the operating airline name (as returned by the API).
So in this case, we give weighting to the normal airline name (as it
is formatted in a much prettier format).  If this assumption turns
out not to be true, we will need to change the handling to associate
a particular name with every single FlightCode.