Refactored code to make stack management easier
I'm going to be changing the code to manually manage the stack, so I pulled
all the code related to stack management into one area of the code.

I also improved the way FlightListFragment is created; there is a new one
for each leg of the flight, which makes management easier (no more
timing/recreation issues).