Separate launch hotel data and hotels hotel data
We want to set the current location/tonight hotel search response to Db (to be
used in PhoneSearchActivity) only if there does not exist any results already.

Store the launch properties in it's own class to separate the data such that
when the user return to the launch screen after performing a new hotel search,
they still have the properties from the current location/tonight search.