Prevent click events from passing to slaveView
when exchanging MotionEvents to scroll the two ListViews in unison, we want to
prevent click events from being dispatched to the slave view. This attempts
to prevent that from occurring by queueing up the a stream of MotionEvents
which starts with a ACTION_DOWN and ends with an ACTION_UP. The queue will
be cleared if it has been determined that a click event occurred.

It was prevented before by sending the CANCEL event to the slave ListView,
although now that there exists a selector, it showed a graphical cue to the
slave view row as selected which we do not want.